backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Celery and Redis
celery==5.3.6
redis==5.0.1

# WebSockets
websockets==12.0
python-socketio==5.11.0

# AI / LLM clients
openai>=1.3.7
groq>=0.9.0
google-generativeai>=0.8.5
elevenlabs>=1.57.0
fish-audio-sdk
together>=1.2.0,<2.0.0
replicate==0.22.0
fal_client==0.4.1
huggingface-hub>=0.23.0

# Media / Audio / Video
moviepy==1.0.3
pydub>=0.25.1
opencv-python==4.10.0.84
numpy==1.26.4
edge-tts==7.2.3
Pillow>=10.2.0

# Transcription / local models
openai-whisper==20231117
torch>=2.0.0
tbb>=2021.11.0

# HTTP and API clients
requests==2.32.3
httpx==0.26.0
aiohttp==3.9.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-magic==0.4.27
reportlab>=4.0.0
psutil
cryptography>=41.0.0
pytz==2023.3

# Storage
boto3==1.34.34  # For S3 storage

# Monitoring and Logging
sentry-sdk==1.39.2
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx-sse==0.4.0

# Security
wmi>=1.5.1; sys_platform == "win32"

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# OAuth
authlib==1.3.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

